refactor: move and extract LessonList and LessonListItem
