+++ src/components/TeacherComponents/LessonListItem/LessonListItem.stories.tsx
@@ -1,9 +1,10 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import AnalyticsDecorator from "../../../../storybook-decorators/AnalyticsDecorator";
-import lessonListingFixture from "../../../../node-lib/curriculum-api/fixtures/lessonListing.fixture";
-
 import LessonListItem, { LessonListItemProps } from "./LessonListItem";
+
+import AnalyticsDecorator from "@/storybook-decorators/AnalyticsDecorator";
+import lessonListingFixture from "@/node-lib/curriculum-api/fixtures/lessonListing.fixture";
+
 
 const lessonListing = lessonListingFixture();
